py/azurecognitiveservice/requirements.txt
azure-cognitiveservices-vision-computervision==0.9.0
azure-common==1.1.28
azure-core==1.28.0
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4
isodate==0.6.1
msrest==0.7.1
oauthlib==3.2.2
Pillow==10.0.0
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
typing_extensions==4.7.1
urllib3==2.0.4
